Re-organize stdlib
Rename net module to pkt and move PcapWriter there. This is a
self-contained library for generating packets and pcap files.

Move tcp4 and udp4 in to ezpkt. This is a library to make it easy to
fill in the fields when generating packets. It provides abstractions
like bi-directional flows.

Remove the dependency on Val and Error in pkt/ezpkt. This leaves stdlib
to be nothing but plumbing, simply hooking up the functionality of pkt
and ezpkt into the resynth interpreter/runtime.