Add "/" operator for joining IP and port
There is the socket literal notation: 8.8.8.8:53, but this doesn't work
when you need to join variables or sub-expressions to create a socket
variable to pass into a function.

We could add a function to do this joining (and maybe we will do that
later), but it would be nice to start adding operators to the language.

We can't use colon for this operator because then the grammar would
become ambiguous when naming arguments to function. So instead we use /
for this:
- ip/port
- 1.2.3.4/port
- ip/1234