Allow U64 and Ip4 to be coerced to strings
We should coerce every type which has an obvious and unambiguous
representation in bytes (ie. all integer types, and IP addresses should
be represented as big-endian 2's complement.

We fix the DNS answer API to now take Str instead of Ip4 so that we can
produce arbitrary DNS answers which aren't just for A records.

Finally, fix the DHCP example to use 0.0.0.0 in place of "|00 00 00 00|"
which is more easily readable.