Allow setting frag_off on UDP datagrams / TCP segs
This isn't a convincing implementation of fragmentation it merely allows
us to set the frag offset field.

The only purpose of this is to be able to easily craft a specific type
of test where we have a packet that perfectly well interpretable as a
TCP or UDP message, but it shouldn't be interpreted that way because the
fragment offset is non-zero.

To actually implement fragmentation you need to account for the fact
that the TCP/UDP header itself is part of the data to be fragmented and
there's currently no way to deal with that.